Feature/auth

+++ src/main/java/com/va1err/IssueTracker/services/UserService.java
@@ -4,6 +4,10 @@ import com.va1err.IssueTracker.dto.requests.UserRequest;
 import com.va1err.IssueTracker.dto.responses.IssueResponse;
 import com.va1err.IssueTracker.dto.responses.ProjectResponse;
 import com.va1err.IssueTracker.dto.responses.UserResponse;
+import com.va1err.IssueTracker.enums.Role;
+import com.va1err.IssueTracker.enums.Priority;
+import com.va1err.IssueTracker.enums.Role;
+import com.va1err.IssueTracker.enums.Status;
 import com.va1err.IssueTracker.exceptions.IssueNotFoundException;
 import com.va1err.IssueTracker.exceptions.ProjectNotFoundException;
 import com.va1err.IssueTracker.exceptions.UserNotFoundException;
@@ -16,8 +20,11 @@ import com.va1err.IssueTracker.repositories.UserRepository;
 import com.va1err.IssueTracker.utils.IssueUtil;
 import com.va1err.IssueTracker.utils.ProjectUtil;
 import com.va1err.IssueTracker.utils.UserUtil;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import java.nio.file.AccessDeniedException;
 import java.util.List;
 
 @Service
@@ -46,10 +53,15 @@ public class UserService {
         return UserUtil.toResponse(user);
     }
 
-    public List<UserResponse> getAllUsers() {
-        List<User> users = userRepository.findAll();
+    public Page<UserResponse> getAllUsers(Role role, Pageable pageable) {
+        Page<User> page;
 
-        return users.stream().map(UserUtil::toResponse).toList();
+        if (role == null)
+            page = userRepository.findAll(pageable);
+        else
+            page = userRepository.findAllByRole(role, pageable);
+
+        return page.map(UserUtil::toResponse);
     }
 
     public UserResponse getUserById(Long id) {
@@ -74,16 +86,25 @@ public class UserService {
         return ProjectUtil.toResponse(project);
     }
 
-    public List<IssueResponse> getUserProjectIssues(Long userId, Long projectId) {
+    public Page<IssueResponse> getUserProjectIssues(Status status, Priority priority, Pageable pageable, Long userId, Long projectId) {
         if (!userRepository.existsById(userId))
             throw new UserNotFoundException(userId);
 
         if (!projectRepository.existsByOwnerIdAndId(userId, projectId))
             throw new ProjectNotFoundException(projectId);
 
-        List<Issue> issues = issueRepository.findAllByOwnerIdAndProjectId(userId, projectId);
+        Page<Issue> page;
 
-        return issues.stream().map(IssueUtil::toResponse).toList();
+        if (status == null && priority == null)
+            page = issueRepository.findAllByOwnerIdAndProjectId(pageable, userId, projectId);
+        else if (status != null && priority == null)
+            page = issueRepository.findAllByOwnerIdAndProjectIdAndStatus(pageable, status, userId, projectId);
+        else if (status == null)
+            page = issueRepository.findAllByOwnerIdAndProjectIdAndPriority(pageable, priority, userId, projectId);
+        else
+            page = issueRepository.findAllByOwnerIdAndProjectIdAndStatusAndPriority(pageable, status, priority, userId, projectId);
+
+        return page.map(IssueUtil::toResponse);
     }
 
     public IssueResponse getUserProjectIssueById(Long userId, Long projectId, Long issueId) {
@@ -99,16 +120,28 @@ public class UserService {
         return IssueUtil.toResponse(issue);
     }
 
-    public void deleteUserById(Long id) {
+    public void deleteUserById(Long id, User currentUser) throws AccessDeniedException {
         if (!userRepository.existsById(id))
             throw new UserNotFoundException(id);
+
+        boolean isThemself = id.equals(currentUser.getId());
+        boolean isAdmin = currentUser.getRole() == Role.ADMIN;
+
+        if (!isThemself && !isAdmin)
+            throw new AccessDeniedException("Not allowed to delete other user");
 
         userRepository.deleteById(id);
     }
 
-    public UserResponse updateUserById(Long id, UserRequest request) {
+    public UserResponse updateUserById(Long id, UserRequest request, User currentUser) throws AccessDeniedException {
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new UserNotFoundException(id));
+
+        boolean isThemself = id.equals(currentUser.getId());
+        boolean isAdmin = currentUser.getRole() == Role.ADMIN;
+
+        if (!isThemself && !isAdmin)
+            throw new AccessDeniedException("Not allowed to update other user");
 
         user.setEmail(request.getEmail());
         user.setUsername(request.getUsername());
